chapter_001/src/test/java/ru/job4j/array/SortSelectedTest.java
package ru.job4j.array;

import org.junit.Test;

import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;

/**
 * Тест класс сортировки массива
 */
public class SortSelectedTest {
    @Test
    public void whenSort() {
        int[] input = new int[]{3, 4, 1, 2, 5};
        int result[] = SortSelected.sort(input);
        int[] expect = new int[]{1, 2, 3, 4, 5};
        assertThat(result, is(expect));
    }

    @Test
    public void whenSort2() {
        int[] input = new int[]{4, 2, 0};
        int[] result = SortSelected.sort(input);
        int[] expect = new int[]{0, 2, 4};
        assertThat(result, is(expect));
    }
}